Added WebApp folder where the function is made into an App
