+++ WebApp/templates/index.html
@@ -1,0 +1,87 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+<!--  -->
+		<style>
+			body, html {
+			    margin: 0;
+			    font-family: sans-serif;
+			    background-color: lightblue;
+			}
+
+			.content {
+			    margin:  auto;
+			    width: 400px;
+			}
+
+			div {
+			  	border: 0px solid black;
+			  	margin: 25px 50px 75px 100px;
+			  	background-color: lightblue;
+			}
+
+			table, td, th {
+			    border: 1px solid #aaa;
+			}
+
+			table {
+			    border-collapse: collapse;
+			    width: 50%;
+			}
+			table.center {
+  				margin-left: auto;
+  				margin-right: auto;
+			}
+
+			th {
+			    height: 30px;
+			}
+
+			td {
+			    text-align: center;
+			    padding: 5px;
+			}
+
+			.form {
+			    margin-top: 20px;
+			}
+
+			#content {
+			    width: 50%;
+			}
+		</style>
+<!--  -->
+		<title>Addition</title>
+</head>
+
+<!--  -->
+<!--  -->
+
+<body>
+	<div>
+	<form action="/predict" method="post">
+		<table class="center">
+				<tr>
+					<td>Load Data (has to be a URL)</td>
+					<td>Select Model</td>
+				</tr>
+<!--  -->
+        <tr>
+          <td><input name="file"/></td>
+          <td>
+            <select name="comp_select" class="selectpicker form-control">
+              {% for m in models %}
+              <option value="{{ m.name }}">{{ m.name }}</option>
+              {% endfor %}
+            </select>
+          </td>
+        </tr>
+    </table>
+		<button type="submit">Predict</button>
+	</form>
+	</div>
+</body>
+</html>
