+++ templates/predict.html
@@ -1,15 +1,15 @@
 {% extends 'base.html' %}
 
 {% block head %}
-<title>Task Master</title>
+<title>Predicting</title>
 {% endblock %}
 
 {% block body %}
 <div class="content">
-    <h1 style="text-align: center">Update Task</h1>
+    <h1 style="text-align: center">Prediction</h1>
 
     <div class="form">
-        <form action="/update/{{task.id}}" method="POST">
+        <form action="/predict/{{task.id}}" method="POST">
             <input type="text" name="content" id="content" value="{{task.content}}">
             <input type="submit" value="Update">
         </form>
