+++ templates/train.html
@@ -1,0 +1,18 @@
+{% extends 'base.html' %}
+
+{% block head %}
+<title>Training</title>
+{% endblock %}
+
+{% block body %}
+<div class="content">
+    <h1 style="text-align: center">Training Models</h1>
+
+    <div class="form">
+        <form action="/train" method="POST">
+            <input type="text" name="content" id="content" value="XXXXXX">
+            <input type="submit" value="Update">
+        </form>
+    </div>
+</div>
+{% endblock %}
